non aggressive tests work

+++ localtest.sh
@@ -5,9 +5,9 @@ fi
 docker images alpine | grep alpine > /dev/null || docker pull alpine
 docker build -t droutertest -f $BASEDIR/Dockertest $BASEDIR
 echo "$@"
-docker run -it --name=drntest_drouter --privileged --rm --pid=host -v /var/run/docker.sock:/var/run/docker.sock -v $BASEDIR/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -coverprofile=/coverage/cover.html $@"
+docker run -it --name=drntest_drouter --privileged --rm -e TERM=xterm --pid=host -v /var/run/docker.sock:/var/run/docker.sock -v $BASEDIR/coverage:/coverage droutertest "go test github.com/TrilliumIT/docker-drouter/drouter -coverprofile=/coverage/cover.html $@"
 ec=$?
-docker rmi $(docker images -q -f dangling=true)
+[ -z $(docker images -q -f dangling=true) ] || docker rmi $(docker images -q -f dangling=true)
 if [ -e $BASEDIR/coverage/cover.html ]; then
 	echo "Press enter to view coverage"
 	read
